moved entities into entity
