+++ entities/typos.py
