+++ entities/vocab.py
@@ -4,12 +4,12 @@ from __future__ import annotations ## used for cheating the circular import issu
 import typing
 
 ## custom modules
-from modules.typos import typo as typo_blueprint
-from modules.typos import incorrectTypo as incorrect_typo_blueprint
+from entities.typos import typo as typo_blueprint
+from entities.typos import incorrectTypo as incorrect_typo_blueprint
 
-from modules.csep import csep as csep
+from entities.csep import csep as csep
 
-from modules import words
+from entities import words
 
 if(typing.TYPE_CHECKING): ## used for cheating the circular import issue that occurs when i need to type check some things
     from handlers.localHandler import localHandler
